requirement.txt
tensorflow
fastapi
uvicorn
numpy
matplotlib
pillow
python-multipart
jinja2